refactor: detail페이지 component 분리
